UPSTREAM: apparmor: make signal label match work when matching stacked labels
Given a label with a profile stack of
    A//&B or A//&C ...

A ptrace rule should be able to specify a generic trace pattern with
a rule like

    signal send A//&**,

however this is failing because while the correct label match routine
is called, it is being done post label decomposition so it is always
being done against a profile instead of the stacked label.

To fix this refactor the cross check to pass the full peer label in to
the label_match.

Change-Id: I371c8a628454b05fc9859dc05e3eb51138868a2a
Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>